Optimizers

This package provides a set optimizers for real-valued functions of a number of complex variables, $f:\mathbb{C}^N\to\mathbb{R}$, which we subdivide into three categories

First-order optimizers

To reach a local minimum of $f(\bm z)$ with $\bm z \in \mathbb{C}^N$, this methods start from some seed value of $\bm z = \bm z^0$ and iterate over $k$ performing a first-order update,

\[\bm z^{k+1} = \bm z^k - a^k \bm g^{k}(\bm z^k, b^k),\]

where $\bm g^k(\bm z^k, b^k)$ is a randomly directed finite-difference estimate of the gradient of $f$, $\partial f / \partial\bm{z}$, and

\[\begin{aligned} a^k &= \frac{a}{(k + A + 1)^s}, \\ b^k &= \frac{b}{(k + 1)^t}, \end{aligned}\]

are convergence factors determined by the input gain parameters $a$, $b$, $A$, $s$, and $t$.

By default, standard gains are defined and used package-wide from the dictionary ComplexSPSA.gains,

ComplexSPSA.gains — Constant

gains = Dict(
    :a => 3.0,
    :b => 0.1,
    :A => 0.0,
    :s => 0.602,
    :t => 0.101,
)

Contains the gain parameters used for the optimizers defined within the ComplexSPSA module. By default, the standard gains are used.

where they can be changed on run-time, and the optimizers will use their value as defined at the moment when they are called.

SPSA on complex

In particular, two first-order methods are implemented. The first, called SPSA_on_complex, is based on the SPSA optimization method for real variables of real functions, and works by treating each complex variable as a pair of real variables. This method is commonly used for the minimization of real functions of complex variables, but may result cumbersome on domains where the complex algebra is natively involved.

ComplexSPSA.SPSA_on_complex — Function

SPSA_on_complex(f::Function, z₀::Vector, Niters = 200;
                sign = -1,
                Ncalibrate = 0,
                initial_iteration = 1,
                constant_learning_rate = false,
                a = gains[:a], b = gains[:b],
                A = gains[:A], s = gains[:s], t = gains[:t],
                )

The Simultaneous Perturbation Stochastic Approximation (SPSA) optimizer is a stochastic method for optimizing real functions of a number of real variables.

This function performs SPSA optimization of the real-valued function f of complex variables by treating each complex varible as a pair of real variables, starting from the complex vector z₀ and iterating Niter times. Then, returns a complex matrix, zacc, with size (length(z₀), Niters), such that zacc[i, j] corresponds to the value of the i-th complex variable on the j-th iteration.

The input parameters a, b, A, s, and t can be provided as keyword arguments of the function. If they are not provided explicitly, they are selected at runtime from the ComplexSPSA.gains dictionary.

Automatic calibration for the input parameter a can be accomplished taking Ncalibrate samples using the method defined on Kandala et. al. (2017), Sec. 11. By default, the calibration is disabled (Ncalibrate = 0).

CSPSA

A second method called CSPSA is also provided, which has the advantage of having being formulated in terms of complex variables, thus resulting more natural in natively-complex domains.

ComplexSPSA.CSPSA — Function

CSPSA(f::Function, z₀::Vector, Niters = 200;
      sign = -1,
      Ncalibrate = 0,
      initial_iteration = 1,
      constant_learning_rate = false,
      a = gains[:a], b = gains[:b],
      A = gains[:A], s = gains[:s], t = gains[:t],
      )

The Complex Simultaneous Perturbation Stochastic Approximation (CSPSA) optimizer is a method for optimizing real functions of a number of complex variables, and corresponds to a complex generalization of the SPSA method.

This function performs CSPSA optimization of the real-valued function f, starting from a vector of complex variables z₀ and iterating Niter times. Then, returns a complex matrix, zacc, with size (length(z₀), Niters), such that zacc[i, j] corresponds to the value of the i-th complex variable on the j-th iteration.

The input parameters a, b, A, s, and t can be provided as keyword arguments of the function. If they are not provided explicitly, they are selected at runtime from the ComplexSPSA.gains dictionary.

Automatic calibration for the input parameter a can be accomplished taking Ncalibrate samples using the method defined on Kandala et. al. (2017), Sec. 11. By default, the calibration is disabled (Ncalibrate = 0).

Second-order optimizers

ComplexSPSA.SPSA2_on_complex — Function

SPSA2_on_complex(f::Function, z₀::Vector, Niters = 200;
                 sign = -1,
                 hessian_delay = 0,
                 initial_iteration = 1,
                 constant_learning_rate = false,
                 a = gains[:a], b = gains[:b],
                 A = gains[:A], s = gains[:s], t = gains[:t],
                 )

The second-order SPSA, commonly referred to as 2-SPSA method is a second-order stochastic optimization method based on SPSA, which additional to a gradient estimate performs a Hessian correction on the update rule to optimize real-valued functions of a number of real variables.

This function performs second-order SPSA optimization of the real-valued function f of complex variables by treating each complex variable as a pair of real variables, starting from the complex vector z₀ and iterating Niter times. Then, returns a complex matrix, zacc, with size (length(z₀), Niters), such that zacc[i, j] corresponds to the value of the i-th complex variable on the j-th iteration.

The input parameters a, b, A, s, and t can be provided as keyword arguments of the function. If not provided explicitly, they are selected at runtime from the ComplexSPSA.gains dictionary.

Since second-order effects usually show improvements once the seed value is closer to a local minimum, it is possible to accept a number hessian_delay of first-order iterations before including the application of the Hessian information.

Notes

The value of a is only required to perform a possible number of initial first-order iterations (via hessian_delay), since the second-order iterations yield an optimum value for a = 1.

ComplexSPSA.CSPSA2 — Function

CSPSA2(f::Function, z₀::Vector, Niters = 200;
       sign = -1,
       hessian_delay = 0,
       initial_iteration = 1,
       constant_learning_rate = false,
       a = gains[:a], b = gains[:b],
       A = gains[:A], s = gains[:s], t = gains[:t],
       )

The second-order CSPSA method, CSPSA2, is a second-order stochastic optimization method based on CSPSA, which additional to a gradient estimate performs a Hessian correction on the update rule to optimize real-valued functions of a number of complex variables.

This function performs second-order CSPSA optimization of the real-valued function f of complex variables, starting from the complex vector z₀ and iterating Niter times. Then, returns a complex matrix, zacc, with size (length(z₀), Niters), such that zacc[i, j] corresponds to the value of the i-th complex variable on the j-th iteration.

The input parameters a, b, A, s, and t can be provided as keyword arguments of the function. If they are not provided explicitly, they are selected at runtime from the ComplexSPSA.gains dictionary.

Since second-order effects usually show improvements once the seed value is closer to a local minimum, it is possible to accept a number hessian_delay of first-order iterations before including the application of the Hessian information.

Notes

The value of a is only required to perform a possible number of initial first-order iterations (via hessian_delay), since the second-order iterations yield an optimum value for a = 1.

Natural gradient

ComplexSPSA.SPSA_QN_on_complex — Function

SPSA_QN_on_complex(f::Function, z₀::Vector, Niters = 200;
                   sign = -1,
                   hessian_delay = 0,
                   initial_iteration = 1,
                   constant_learning_rate = false,
                   a = gains[:a], b = gains[:b],
                   A = gains[:A], s = gains[:s], t = gains[:t],
                   )

The Quantum Natural SPSA, presented by Gacon et. al. (2021), is a second-order stochastic optimization method based on 2-SPSA, where the second-order correction comes from the Hessian of the Fubini-Study metric of the problem instead of the Hessian of the function under optimization.

Note that the metric must be a function taking two input vectors, and returning minus a half of the fidelity between the states generated each from one input,

$ \text{metric}(\vec z₁, \vec z₂) = -\frac{1}{2} |\langle ψ(\vec z₁) | ψ(\vec z₂) \rangle|^2, $

where $ ψ(\vec z) $ is the quantum state parameterized with the variables $ \vec z $.

This function performs Quantum Natural SPSA optimization of the real-valued function f of complex variables by treating each complex variable as a pair of real variables, starting from the complex vector z₀ and iterating Niter times. Then, returns a complex matrix, zacc, with size (length(z₀), Niters), such that zacc[i, j] corresponds to the value of the i-th complex variable on the j-th iteration.

The input parameters a, b, A, s, and t can be provided as keyword arguments of the function. If not provided explicitly, they are selected at runtime from the ComplexSPSA.gains dictionary.

Since second-order effects usually show improvements once the seed value is closer to a local minimum, it is possible to accept a number hessian_delay of first-order iterations before including the application of the Hessian information.

Notes

The value of a is only required to perform a possible number of initial first-order iterations (via hessian_delay), since the second-order iterations yield an optimum value for a = 1.

ComplexSPSA.CSPSA_QN — Function

CSPSA_QN(f::Function, metric::Function, z₀::Vector, Niters = 200;
         sign = -1,
         hessian_delay = 0,
         initial_iteration = 1,
         constant_learning_rate = false,
         a = gains[:a], b = gains[:b],
         A = gains[:A], s = gains[:s], t = gains[:t],
         )

The Quantum Natural CSPSA (QN-CSPSA), is a second-order stochastic optimization method which, analogous to the Quantum Natural SPSA by Gacon et. al. (2021), takes into account a stochastic approximation of the Fubiny-Study metric instead of the usual Hessian correction from CSPSA2. However, the main difference between QN-CSPSA and QN-SPSA is that the former is natively formulated in terms of complex variables, while the latter requires real variables. Note that the metric must be a function taking two input vectors, and returning minus a half of the fidelity between the states generated each from one input,

$ \text{metric}(\vec z₁, \vec z₂) = -\frac{1}{2} |\langle ψ(\vec z₁) | ψ(\vec z₂) \rangle|^2, $

where $ ψ(\vec z) $ is the quantum state parameterized with the variables $ \vec z $.

This function performs Quantum Natural CSPSA optimization of the real-valued function f of complex variables, starting from the complex vector z₀ and iterating Niter times. Then, returns a complex matrix, zacc, with size (length(z₀), Niters), such that zacc[i, j] corresponds to the value of the i-th complex variable on the j-th iteration.

The input parameters a, b, A, s, and t can be provided as keyword arguments of the function. If not provided explicitly, they are selected at runtime from the ComplexSPSA.gains dictionary.

Since second-order effects usually show improvements once the seed value is closer to a local minimum, it is possible to accept a number hessian_delay of first-order iterations before including the application of the Hessian information.

Notes

The value of a is only required to perform a possible number of initial first-order iterations (via hessian_delay), since the second-order iterations yield an optimum value for a = 1.

Missing docstring.

Missing docstring for SPSA_QN_scalar_on_complex. Check Documenter's build log for details.

Missing docstring.

Missing docstring for CSPSA_QN_scalar. Check Documenter's build log for details.