Qiskit Wrapper

For purposes of ease of comparison, the ComplexSPSA module includes a Qiskit submodule which wraps some SPSA-based optimizers from the python library Qiskit, and exposes them with an interface common to the rest of the module. The wrappers use the default options as defined by Qiskit for each optimizer.

Index of optimizers implemented

ComplexSPSA.Qiskit.SPSA

ComplexSPSA.Qiskit.SPSA2

ComplexSPSA.Qiskit.SPSA2_on_complex

ComplexSPSA.Qiskit.SPSA_NG

ComplexSPSA.Qiskit.SPSA_NG_on_complex

ComplexSPSA.Qiskit.SPSA_on_complex

Ensure julia can find Qiskit

To make sure that julia can find and use the Qiskit library, you can execute the function

ComplexSPSA.Qiskit.pip_install_dependencies — Function

pip_install_dependencies()

Ensure the Qiskit library is installed on the python distribution reached by julia.

Notes

Assumes the pip tool is present on the system.

which is provided by the Qiskit submodule, and may thus be executed as

using ComplexSPSA
ComplexSPSA.Qiskit.pip_install_dependencies()

Using the optimizers

To expose the methods defined on the Qiskit submodule, you can run

using ComplexSPSA.Qiskit

and then, use the optimizers as

x = Qiskit.SPSA(f, guess, Niters)

assuming you have previously defined a valid function, f, initial array of parameters, guess, and number of iterations, Niters.