Real-variable optimizers

Simple, thin wrappers around the SPSA-based optimizers implemented by Qiskit. This optimizers only work on real functions of real variables.

ComplexSPSA.Qiskit.SPSA — Method

SPSA(f, guess, Niters;
     a = Qiskit.gains[:a], b = Qiskit.gains[:b],
     A = Qiskit.gains[:A], s = Qiskit.gains[:s], t = Qiskit.gains[:t] )

Wrapper around the first order SPSA optimizer from Qiskit.

ComplexSPSA.Qiskit.SPSA2 — Method

SPSA2(f, guess, Niters;
      b = Qiskit.gains[:b], A = Qiskit.gains[:A],
      s = Qiskit.gains[:s], t = Qiskit.gains[:t] )

Wrapper around the second order SPSA optimizer from Qiskit.

ComplexSPSA.Qiskit.SPSA_NG — Method

SPSA_NG(f, fidelity, guess, Niters;
        b = Qiskit.gains[:b], A = Qiskit.gains[:A],
        s = Qiskit.gains[:s], t = Qiskit.gains[:t] )

Wrapper around the QNSPSA optimizer from Qiskit.

Complex-variable optimizers

Optimizers which separate the optimization problem of a real function of $N$ complex variables as the equivalent problem of optimizing a real function of $2N$ real variables, and then solve it by means of the Real-variable optimizers.

ComplexSPSA.Qiskit.SPSA2_on_complex — Method

SPSA2_on_complex(f, guess::Vector{Complex{Float64}}, Niters;
                 b = Qiskit.gains[:b], A = Qiskit.gains[:A],
                 s = Qiskit.gains[:s], t = Qiskit.gains[:t] )

Wrapper around SPSA2 which internally converts the complex-valued initial point, guess, and objective function, f, to be optimized as a real-variable problem with twice the parameters.

ComplexSPSA.Qiskit.SPSA_NG_on_complex — Method

SPSA_NG_on_complex(f, fidelity, guess::Vector{Complex{Float64}}, Niters;
                   b = Qiskit.gains[:b], A = Qiskit.gains[:A],
                   s = Qiskit.gains[:s], t = Qiskit.gains[:t] )

Wrapper around SPSA_NG which internally converts the complex-valued initial point, guess, the fidelity, and the objective function, f, to be optimized as a real-variable problem with twice the parameters.

ComplexSPSA.Qiskit.SPSA_on_complex — Method

SPSA_on_complex(f, guess::Vector{Complex{Float64}}, Niters;
                a = Qiskit.gains[:a], b = Qiskit.gains[:b],
                A = Qiskit.gains[:A], s = Qiskit.gains[:s], t = Qiskit.gains[:t] )

Wrapper around SPSA which internally converts the complex-valued initial point, guess, and objective function, f, to be optimized as a real-variable problem with twice the parameters.