Tools

Some convenience functions are exposed to the user to solve simple, usual tasks that may appear commonly on the use-cases of this module. Those tools are listed below.

Index

ComplexSPSA.apply_along_dim

ComplexSPSA.simulate_experiment

ComplexSPSA.squeeze

Documentation

ComplexSPSA.apply_along_dim — Method

apply_along_dim(f::Function, A::AbstractArray; dim::Integer = 1)

Apply a function, f, over slices along a dimension, dim, of a multidimensional array, A.

Example

julia> a = [ 1 2; 3 4 ]
2×2 Matrix{Int64}:
 1  2
 3  4

julia> apply_along_dim(sum, a, dim = 1)
1×2 Matrix{Int64}:
 4  6

ComplexSPSA.simulate_experiment — Function

simulate_experiment(refvalue, Nmeasures = Inf)

Simulates the experimental measurement with Nmeasures number of tries of an observable whose theoretical value is refvalue.

The experiental result is simulated by sampling a binomial distribution with Nmeasures tries and success rate refvalue, and normalizing the result against the number of tries.

Notes

If the number of measurements, Nmeasures, is infinite, the reference value, refvalue, is returned exactly.

It is assumed that 0.0 <= refvalue <= 1.0. If not, refvalue will be taken as its closest boundary.

ComplexSPSA.squeeze — Method

squeeze(A :: AbstractArray)

Remove the singleton (length-1) dimensions of a multidimensional array.

Example

julia> a = rand(10, 3, 1, 4);

julia> size(a)
(10, 3, 1, 4)

julia> size( squeeze(a) )
(10, 3, 4)